Add `:propagate_per_column` option to propagate changes to future versions
The option has to be set per class on:

```
class Demo < Sequel::Model
  plugin :bitemporal, version_class: EmployeeVersion,
                      propagate_per_column: true
end
```

The goal is to allow changes to individual column to be propagates to
future versions. To be propagated, the future version column should have the
same value as its previous version before update and the value should
have changed during the current update.

Note: This commit does not include audit update